20221129complete

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jingchen <jingchen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/29 17:27:39 by jingchen          #+#    #+#             */
-/*   Updated: 2022/11/29 17:29:53 by jingchen         ###   ########.fr       */
+/*   Updated: 2022/11/29 18:46:31 by jingchen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,10 +21,10 @@
 int	ft_printf(char const *format, ...);
 int	ft_putchar(int c);
 int	ft_putstr(char *s);
-int	ft_putnbr(long int n);
+int	ft_putnbr(int n);
 int	ft_putunbr(unsigned int n);
 int	ft_putptr(unsigned long n);
-int	ft_puthexupper(unsigned long n);
-int	ft_puthexlower(unsigned long n);
+int	ft_puthexupper(unsigned int n);
+int	ft_puthexlower(unsigned int n);
 
 #endif
